+++ config/initializers/kaminari_config.rb
@@ -1,6 +1,6 @@
 Kaminari.configure do |config|
   config.default_per_page = 24
-  config.window = 4
+  config.window = 3
   # config.outer_window = 2
   # config.left = 0
   # config.right = 0